Code cleanup and refactor
Allow the user to interrupt with CTRL+C. We will still cleanup and output metrics if requested.
Proper capitalization on types, variables, and functions.
Add a function to sort latency data, and store it once. This should make metrics calculation faster.
Fix a bug when calculating the total throughput for a target file.
Add a helper function to sort latency data, and prevent repeated sorts.