Cleanup Allocate functions
Use os.Truncate instead of creating a file, then truncating it in non-Linux environments.
Return the value of the truncate error since it's the value we're looking for anyway.
In Linux environments, attempt to use the existing file, or Truncate it if a smaller size has been requested. This prevents unnecessary churn on multiple runs when using the `-keep` flag.
Use the returned value for Allocate in Linux environments rather than explicitly returning `nil` on success.