Optimizations and fixes
Create a random source for each routine rather than relying on the global rand. This improves performance for concurrent access.

Don't use the total accumulated copied data as the addition to dataReader.position. Instead, increment by the last copied amount.

Collapse the Random and Repeat cases in reader and writer workers to simplify the logic.

When Seek() fails, return from the worker.